+++ packages/widget-v2/src/components/Button.tsx
@@ -1,0 +1,10 @@
+import { styled } from 'styled-components';
+import { FlexProps, flexProps } from './Layout';
+
+export const Button = styled.button<FlexProps>`
+  all: unset;
+  &:hover {
+    cursor: pointer;
+  }
+  ${flexProps};
+`;