+++ packages/widget/src/hooks/use-account.ts
@@ -14,6 +14,7 @@ import {
   isReadyToCheckLedger,
 } from '../utils/wallet';
 import { useChainByID } from './use-chains';
+import { chainIdToName } from '../chains';
 
 export function useAccount(chainID?: string) {
   const { data: chain } = useChainByID(chainID);
@@ -23,9 +24,11 @@ export function useAccount(chainID?: string) {
 
   const cosmosWallet = useMemo(() => {
     if (chain?.chainType !== 'cosmos') return;
-    const { wallets } = getWalletRepo(chain.chainName);
+    if (chain.chainID.includes('penumbra')) return;
+    const { wallets } = getWalletRepo(chainIdToName(chain.chainID));
     return wallets.find((w) => w.walletName === trackedWallet?.walletName);
   }, [
+    chain?.chainID,
     chain?.chainName,
     chain?.chainType,
     getWalletRepo,