+++ packages/widget/src/index.ts
@@ -1,4 +1,4 @@
-export { SwapWidgetProvider } from './provider';
+export { SwapWidgetProvider, SwapWidgetProviderProps } from './provider';
 export {
   SwapWidget,
   SwapWidgetProps,
@@ -8,3 +8,4 @@ export {
 export { initializeSwapWidget } from './ui/WebComponent';
 export { useAssets } from './provider/assets';
 export { useChains, useChainByID } from './hooks/use-chains';
+export { MinimalWallet } from './hooks/use-make-wallets';