+++ packages/widget/src/provider/skip-provider.tsx
@@ -5,17 +5,18 @@ import { getWalletClient } from '@wagmi/core';
 import { createContext, ReactNode } from 'react';
 import { WalletClient } from 'viem';
 
-import { chainIdToName } from '../chains';
 import { config } from '../lib/wagmi';
 import { trackWallet } from '../store/track-wallet';
 import { gracefullyConnect, isWalletClientUsingLedger } from '../utils/wallet';
-import { WidgetConfig } from '.';
+import { chainIdToName } from '../chains';
+import { MinimalWallet } from '../hooks/use-make-wallets';
 
 export const SkipContext = createContext<
   | {
       skipClient: SkipRouter;
       apiURL?: string;
       endpointOptions?: SkipRouterOptions['endpointOptions'];
+      makeDestinationWallets?: (chainID: string) => MinimalWallet[];
     }
   | undefined
 >(undefined);
@@ -24,17 +25,19 @@ export function SkipProvider({
   children,
   apiURL,
   endpointOptions,
+  makeDestinationWallets,
 }: {
   children: ReactNode;
   apiURL?: string;
   endpointOptions?: SkipRouterOptions['endpointOptions'];
+  makeDestinationWallets?: (chainID: string) => MinimalWallet[];
 }) {
   const { getWalletRepo } = useManager();
   const { wallets } = useWallet();
 
   const skipClient = new SkipRouter({
     getCosmosSigner: async (chainID) => {
-      const chainName = chainIdToName[chainID];
+      const chainName = chainIdToName(chainID);
       if (!chainName) {
         throw new Error(`getCosmosSigner error: unknown chainID '${chainID}'`);
       }
@@ -109,6 +112,7 @@ export function SkipProvider({
         skipClient,
         apiURL,
         endpointOptions,
+        makeDestinationWallets,
       }}
     >
       {children}